separate await-until and move to src/utils
await-until becomes useful in streamrChannel, so better move it to src
