+++ test/system/data-union-demo.js
@@ -14,7 +14,7 @@ const {
 const StreamrChannel = require("../../src/streamrChannel")
 
 const sleep = require("../../src/utils/sleep-promise")
-const { untilStreamContains, untilStreamMatches, capture } = require("../utils/await-until")
+const { untilStreamContains, untilStreamMatches, capture } = require("../utils/await-process-stream")
 const deployContract = require("../../src/utils/deployContract")
 
 const ERC20Mintable = require("../../build/ERC20Mintable.json")
