+++ test/system/demo-using-client.js
@@ -14,7 +14,7 @@ const {
 } = require("ethers")
 
 const sleep = require("../../src/utils/sleep-promise")
-const { untilStreamContains } = require("../utils/await-until")
+const { untilStreamContains } = require("../utils/await-process-stream")
 
 const ERC20Mintable = require("../../build/ERC20Mintable.json")
 const DataUnion = require("../../build/DataunionVault")
